Fix: Add `userId` parameter for task models

+++ src/model/task/deleteTask.ts
@@ -3,13 +3,14 @@ import db from "model";
 import type { ResultSetHeader } from "mysql2";
 
 interface DeleteTaskProps {
+  userId: number;
   id: number;
 }
 
-export default async function deleteTask({ id }: DeleteTaskProps) {
+export default async function deleteTask({ userId, id }: DeleteTaskProps) {
   const queryResult = await db.query<ResultSetHeader>(
-    "DELETE FROM task WHERE id = ?",
-    [id]
+    "DELETE FROM task WHERE creator = ? AND id = ?",
+    [userId, id]
   );
 
   return queryResult;
